third_party/WebKit/LayoutTests/accessibility/aria-controls-with-tabs-expected.txt

Crust
Veges
Test
Select Crust
Select Crust

This tests that the aria tab item becomes selected if either aria-selected is used, or if aria-controls points to an item that contains KB focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tab2.isSelected is true
PASS tab1.isSelected is false
PASS tab2.isSelected is false
PASS tab1.isSelected is true
PASS successfullyParsed is true

TEST COMPLETE
